third_party/WebKit/LayoutTests/external/wpt/html/dom/dynamic-markup-insertion/opening-the-input-stream/document.open-02-expected.txt

This is a testharness.js-based test.
FAIL document.open should redirect to window.open when called with three arguments assert_unreached: The call should be redirected to the real window.open Reached unreachable code
PASS document.open should throw when it has no window and is called with three arguments
Harness: the test ran to completion.
